move test_attribute_io tests to test_attribute.py
